ppo/requirements.txt
numpy>=1.21.0
torch>=1.9.0
gymnasium>=0.28.0
matplotlib>=3.4.0
seaborn>=0.11.0
tensorboard>=2.6.0
pandas>=1.3.0
scipy>=1.7.0
tqdm>=4.62.0
ruamel.yaml==0.19.1